+++ src/components/Game.tsx
@@ -1,10 +1,11 @@
 import { lazy, Suspense, useEffect, useRef, useState } from "react";
 import { GlobeMethods } from "react-globe.gl";
 import { Country } from "../lib/country";
-import { answerName } from "../util/answer";
+import { answerCountry, answerName } from "../util/answer";
 import { useLocalStorage } from "../hooks/useLocalStorage";
 import { Guesses, Stats } from "../lib/localStorage";
 import { dateDiffInDays, today } from "../util/dates";
+import { polygonDistance } from "../util/distance";
 
 const Globe = lazy(() => import("./Globe"));
 const Guesser = lazy(() => import("./Guesser"));
@@ -46,6 +47,7 @@ export default function Game({ reSpin, setShowStats }: Props) {
         return country.properties.NAME === guess;
       });
       if (!foundCountry) throw new Error("Country mapping broken");
+      foundCountry["proximity"] = polygonDistance(foundCountry, answerCountry);
       return foundCountry;
     });
   }