+++ src/components/Footer.tsx
@@ -6,7 +6,7 @@ export default function Footer() {
   const iconWidth = 14;
   const { nightMode } = useContext(ThemeContext).theme;
   return (
-    <footer className="absolute -bottom-24 md:-bottom-36 left-0 py-4 text-xs">
+    <footer className="pt-8 pb-4 text-xs">
       <span className="flex space-x-3">
         <a href="https://the-abe-train.com">by The Abe Train</a>
         <a href="https://twitter.com/theAbeTrain" aria-label="Twitter">
@@ -29,16 +29,6 @@ export default function Footer() {
             <path d={getPath("github")} />
           </svg>
         </a>
-      </span>
-      <span>
-        Learn what qualifies as a &nbsp;
-        <a
-          className="underline"
-          href="https://www.sporcle.com/blog/2013/01/what-is-a-country/"
-        >
-          valid country
-        </a>
-        &nbsp; in this game.
       </span>
     </footer>
   );