+++ src/components/Globe.tsx
@@ -1,4 +1,4 @@
-import { useContext, useEffect, useRef } from "react";
+import { useContext, useEffect, useRef, useState } from "react";
 import ReactGlobe, { GlobeMethods } from "react-globe.gl";
 import { Country } from "../lib/country";
 import { findCentre } from "../util/centre";
@@ -7,6 +7,7 @@ import { turnGlobe } from "../util/turnGlobe";
 import { ThemeContext } from "../context/ThemeContext";
 import { getColour } from "../util/colour";
 import useCheckMobile from "../hooks/useCheckMobile";
+const territoryData: Country[] = require("../data/territories.json").features;
 
 type Props = {
   guesses: Country[];
@@ -14,9 +15,12 @@ type Props = {
 };
 
 export default function Globe({ guesses, globeRef }: Props) {
+  // State
+  const [places, setPlaces] = useState(guesses);
+
   // Theme
-  const { nightMode } = useContext(ThemeContext).theme;
-  const { highContrast } = useContext(ThemeContext).theme;
+  const { nightMode, highContrast } = useContext(ThemeContext).theme;
+
 
   // Check device
   const isMobile = useCheckMobile();
@@ -30,6 +34,17 @@ export default function Globe({ guesses, globeRef }: Props) {
 
   // After each guess
   useEffect(() => {
+    // Add territories to guesses to make shapes
+    const territories: Country[] = [];
+    guesses.forEach((guess) => {
+      const territory = territoryData.find((territory) => {
+        return guess.properties.NAME === territory.properties.SOVEREIGNT;
+      });
+      if (territory) territories.push(territory);
+    });
+    setPlaces(guesses.concat(territories));
+
+    // Turn globe to new spot
     const controls: any = globeRef.current.controls();
     controls.autoRotate = false;
     const newGuess = [...guesses].pop();
@@ -61,24 +76,50 @@ export default function Globe({ guesses, globeRef }: Props) {
     });
   }, [globeRef]);
 
+  // Label colour
+  function getLabel(country: Country) {
+    const name = country.properties.ADMIN;
+    const prox = country.proximity;
+    const dayColour = prox < 750_000 ? "gray-300" : "gray-900";
+    const nightColour = "gray-300";
+    const label = `<b class="text-${dayColour} dark:text-${nightColour}">${name}</b>`;
+    return label;
+  }
+
+  // Polygon altitude
+  function getAltitude(country: Country) {
+    if (!highContrast || country.properties.TYPE === "Territory") return 0.01;
+    const prox = country.proximity;
+    let proxFraction = prox / 2_000_000;
+    proxFraction = Math.min(Math.max(proxFraction, 0.01), 0.95);
+    let alt = (1 - proxFraction) / 10;
+    return alt;
+  }
+
   return (
-    <div ref={containerRef} className="mx-auto cursor-grab" style={extraStyle}>
+    <div
+      ref={containerRef}
+      className="mx-auto cursor-grab text-center"
+      style={extraStyle}
+    >
       <ReactGlobe
         ref={globeRef}
         globeImageUrl={`images/earth-${nightMode ? "night" : "day"}.webp`}
         width={size}
         height={size}
         backgroundColor="#00000000"
-        polygonsData={guesses}
+        polygonsData={places}
+        polygonCapColor={(c) =>
+          // @ts-ignore
+          getColour(c, answerCountry, nightMode, highContrast)
+        }
         // @ts-ignore
-        polygonCapColor={(c) => getColour(c, answerCountry, nightMode, highContrast)}
+        polygonLabel={getLabel}
         // @ts-ignore
-        polygonLabel={({ properties: d }) => `
-        <b class="text-black dark:text-gray-300">${d.ADMIN}</b> 
-        `}
+        polygonAltitude={getAltitude}
+        polygonSideColor="blue"
         onGlobeClick={(d) => turnGlobe(d, globeRef)}
         onPolygonClick={(p, e, c) => turnGlobe(c, globeRef)}
-        polygonSideColor="#00000000"
         polygonStrokeColor="#00000000"
         atmosphereColor={nightMode ? "rgba(63, 201, 255)" : "lightskyblue"}
       />